requirements.txt
# Lightweight Dependencies
python-dotenv==1.0.0
itsdangerous==2.2.0
python-slugify
loguru==0.7.0
python-multipart
pyjwt
bs4
requests
pytest==7.4.0
pytest-asyncio==0.21.0
httpx>=0.28.0,<0.29.0
flake8==6.1.0
black==23.10.0
isort==5.12.0
python-dateutil
fastapi-utilities
lxml
python-docx
fpdf2
XlsxWriter

# Medium-weight Dependencies
fastapi>=0.100.1
uvicorn[standard]==0.23.0
passlib[bcrypt]==1.7.4
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
motor==3.2.0
pymongo==4.4

# Heavyweight Dependencies
langchain===0.3.13
chromadb=== 0.5.0
langchain-community===0.3.13
langchain-chroma===0.1.4
langgraph===0.2.35

slowapi===0.1.9

# Google AI Dependencies
google-generativeai>=0.7.0,<0.8.0
langchain-google-vertexai==2.0.21
langchain-google-genai==2.0.0

unstructured[all-docs]

Pillow
python-dotenv 

torch==2.4.1
torchvision==0.19.1
scikit-learn==1.5.2
kaggle==1.6.17
git+https://github.com/openai/CLIP.git